vnhc-backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
mysqlclient==2.2.4
django-cors-headers==4.4.0
Pillow==10.4.0
python-dotenv==1.0.1
whitenoise==6.7.0
werkzeug==3.0.4
gunicorn==23.0.0
